backend/requirements.txt
fastapi
uvicorn
python-dotenv
bytez
beautifulsoup4
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
google-auth
python-dateuti